activity_api/requirements.txt
altgraph==0.16.1
Click==7.0
Flask==1.0.2
future==0.17.1
itsdangerous==1.1.0
Jinja2==2.10.1
macholib==1.11
MarkupSafe==1.1.1
pefile==2018.8.8
PyInstaller==3.4
pymssql==2.1.4
pywin32-ctypes==0.2.0
PyYAML==5.1
Werkzeug==0.15.2
